refactor(file-upload): -> max filesize 5mb + extracted uploadFiles func
